Added opacity and optical depth
Finally opacity and optical depth are available.

Due to differences in interpolation ant integration algorithms, the results are not exactly the same to CAT. However, CAT doesn´t return exactly those results CO5BOLD is internally computing. The results now are closer to the CO5BOLD results than to CAT results.

The optical depth is computed with the help of the cumulative trapezoid routine from scipy´s integrate submodule. The differences to the optical depth of CAT (cubic integration) is neglegible in the tested case. If significant differences are found, please tell me.

Furthermore some bugfixes and small modifications have been done. It is highly recommended to update downloaded files.