Changed from Qt4 to Qt5
The change of the API might have introduced new bugs, but yet, none has showed up.

The change also brings minor optimizations, especially for interpolation, which is used for the magnetic field-array. Furthermore a log-file has been introduced, which can be used to change the standard directories for opacity-, eos- and model-files. This means, if the corresponding open-dialog is opened, it will be directed to the directory given in the init.log file. It might be used for other quantities as well, later on.