Normalization over time introduced
Now the check-box for Normalization over time works. Checking it, leads
to the computation of the temporal total minimum and maximum value for
every position and direction of plane, e.g. for a box with Nx x Ny x Nz
cells will lead to 2 x Nx x Ny x Nz values computed. By this, the
computation is done once for the current quantity. These values will be
used as minimum and maximum value of the plot. Along time, there won´t
be a change in the value-range plotted for one position and
slice-direction.

However, if the quantity is changed, the values have to be recomputed. A
large time-series loaded, might lead to long times of computation.